+++ src/app/components/file-tree-dialog/file-tree-dialog.component.ts
@@ -2,7 +2,6 @@ import { Component, inject, OnInit } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatTreeModule } from '@angular/material/tree';
-import { LoadingSpinnerComponent } from "../loading-spinner/loading-spinner.component";
 import { CommonModule } from '@angular/common';
 import { MAT_DIALOG_DATA, MatDialogModule } from '@angular/material/dialog';
 import { FileTreeComponent } from '../file-tree/file-tree.component';
@@ -10,7 +9,7 @@ import { FileTreeComponent } from '../file-tree/file-tree.component';
 @Component({
   selector: 'file-tree-dialog',
   imports: [
-    MatTreeModule, MatButtonModule, MatIconModule, LoadingSpinnerComponent, CommonModule, MatDialogModule,
+    MatTreeModule, MatButtonModule, MatIconModule, CommonModule, MatDialogModule,
     FileTreeComponent
   ],
   templateUrl: './file-tree-dialog.component.html',