+++ src/app/pages/publication-collections/publication-collections.component.ts
@@ -4,7 +4,6 @@ import { Component, OnInit } from '@angular/core';
 import { BehaviorSubject, combineLatest, filter, map, Observable, of, switchMap } from 'rxjs';
 import { PublicationCollection } from '../../models/publication';
 import { MatTableModule } from '@angular/material/table';
-import { CustomDatePipe } from '../../pipes/custom-date.pipe';
 import { Column, Deleted } from '../../models/common';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
@@ -25,7 +24,7 @@ import { ConfirmDialogComponent } from '../../components/confirm-dialog/confirm-
 @Component({
   selector: 'publication-collections',
   imports: [
-    CommonModule, MatTableModule, CustomDatePipe, MatIconModule, MatButtonModule, RouterLink, LoadingSpinnerComponent,
+    CommonModule, MatTableModule, MatIconModule, MatButtonModule, RouterLink, LoadingSpinnerComponent,
     PublicationsComponent, MatBadgeModule, CustomTableComponent, CustomTableComponent
   ],
   providers: [DatePipe],